Move translate_duration() into localization file
